[WorkspaceTests] Migrate Workspace tests to new testing infrastructure #2
