+++ Tests/WorkspaceTests/XCTestManifests.swift
@@ -17,7 +17,6 @@ public func allTests() -> [XCTestCaseEntry] {
         testCase(PinsStoreTests.allTests),
         testCase(ToolsVersionWriterTests.allTests),
         testCase(WorkspaceTests.allTests),
-        testCase(WorkspaceTests2.allTests),
     ]
 }
 #endif
